Settings - Wifi notifications: notify when wifi connects [2/2]
Options:
off (default)
a dry white toast
a notification
a notification with a sound

For those of us that survive on wifi islands in small college towns this
is
pretty nice to have

Any notification within 30 seconds of broadcast intent
ACTION_BOOT_COMPLETED
will be canceled to prevent startup notification spamming. 8 (default)
of the
previous connections are remembered and it will update the
notification's content
accordingly if you've restored a connection.

Conflicts:
	res/values/cm_arrays.xml
	res/values/cm_strings.xml

Change-Id: Ica4b868e4db5a77e1ad352b8a2e9308e2429aba0
Signed-off-by: STELIX <ssspinni@gmail.com>