Add customization for Quick Settings [2/2] Allow Quick Settings tiles to be arranged any way you want, delete them or add some.
I removed the "dual" status that was hardcoded into Wifi and BT tiles and added it
as a variable which is true when the tile is on the first row.

Also, the first row is now hardcoded to have only 2 items to follow the same design as before.

Inversion and Hotspot tile are no longer contextual and can be added, rmeoved or rearranged at will.

Change-Id: I9aa405ddfb0bb162f85b2ac8831354a84734a216